Fix file size units in explorer
