+++ mithril-explorer/src/components/Artifacts/SnapshotsList/index.js
@@ -5,21 +5,8 @@ import RawJsonButton from "../../RawJsonButton";
 import { useSelector } from "react-redux";
 import { selectedAggregator } from "../../../store/settingsSlice";
 import LocalDateTime from "../../LocalDateTime";
+import { formatBytes } from "../../../utils";
 
-/*
- * Code from: https://stackoverflow.com/a/18650828
- */
-function formatBytes(bytes, decimals = 2) {
-  if (bytes === 0) return '0 Bytes';
-
-  const k = 1024;
-  const dm = decimals < 0 ? 0 : decimals;
-  const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-
-  const i = Math.floor(Math.log(bytes) / Math.log(k));
-
-  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i];
-}
 
 export default function SnapshotsList(props) {
   const [snapshots, setSnapshots] = useState([]);
